Feature/313 laravel admin image cropping bug (#227)
* fix: add changes to fix filling out the crop data for the file

* fix: changes for solving two bugs

- add change to correct the position of the cropper on the image (when the image has already the crop data in the database)
- change the logic of updating the images. It is needed because the method for updating images deletes the version of the images when it gets the NULL instead of the proper crop data

* feat: change the logic of the is_webp_acceptable function

change the logic of the is_webp_acceptable function so that the function will check the version of Safari. This change allows us to return the images in the WebP format for Safari users.

* Apply prettify formatting changes

---------

+++ app/Http/Controllers/Api/v1/Media/EditImageController.php
@@ -9,7 +9,6 @@ use App\Jobs\RenameFile;
 use App\Model\Admin\MediaFile;
 use App\Model\Admin\MediaFolder;
 use App\Services\Media\MediaFileService;
-use Dompdf\Exception;
 use Illuminate\Http\Response;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Log;
@@ -25,6 +24,7 @@ class EditImageController extends ApiBaseController
         if ($version !== 'original' && empty($versionOptions)) {
             return $this->response(['message' => 'Unknown version'], Response::HTTP_BAD_REQUEST);
         }
+
         $originalFile = MediaFile::findOrFail($request->input('parent_id'));
         if ($originalFile->id === $request->input('id') || $version === 'original') {
             $originalFile->update($request->only(['name', 'alt_text', 'link']));
@@ -41,27 +41,32 @@ class EditImageController extends ApiBaseController
         $fileExt = ($version === 'feed-image') ? 'jpg' : 'webp';
         $path = $mediaFolder->path . '/' . $fileBaseName . '.' . $fileExt;
 
-        $imageWebP = null;
+        if (!Storage::disk('public')->exists($originalFile->path)) {
+            return $this->response(['message' => 'Original File not found'], Response::HTTP_BAD_REQUEST);
+        }
+
         $mediaFile = MediaFile::where('parent_id', $originalFile->id)
             ->where('version', $version)
             ->first();
 
-        if ($mediaFile) {
+        if ($request->filled('crop_data') && $mediaFile) {
             $service->deleteVersionFile($mediaFile);
         }
-        if (!Storage::disk('public')->exists($originalFile->path)) {
-            return $this->response(['message' => 'Original File not found'], Response::HTTP_BAD_REQUEST);
-        }
+
+        $imageWebP = null;
         $cropData = $request->input('crop_data');
+
         if ($cropData) {
             $heightRatio = $request->input('height_ratio');
             $widthRatio = $request->input('width_ratio');
             if (!$heightRatio || !$widthRatio || $heightRatio <= 0 || $widthRatio <= 0) {
                 return $this->response(['message' => 'Bad Request'], Response::HTTP_BAD_REQUEST);
             }
+
             if (empty($cropData['height']) || $cropData['height'] < 1 || empty($cropData['width']) || $cropData['width'] < 1) {
                 return $this->response(['message' => 'Bad Request'], Response::HTTP_BAD_REQUEST);
             }
+
             $imgHeight = intval($cropData['height'] * (1 / (float) $heightRatio));
             $imgWidth = intval($cropData['width'] * (1 / (float) $widthRatio));
             $imgTop = empty($cropData['top']) ? 0 : intval($cropData['top'] * (1 / (float) $heightRatio));
@@ -90,54 +95,41 @@ class EditImageController extends ApiBaseController
 
                 return $this->responseException('Original File not readable', $e);
             }
-        } else {
-            try {
-                $contents = Storage::disk('public')->get($originalFile->path);
-                try {
-                    $imageWebP = Image::make($contents);
-                    $this->storeImage($path, $imageWebP, $fileExt);
-                } catch (\Exception $e) {
-                    Log::error('FileVersion generating failed . ' . $e->getMessage());
-
-                    return $this->responseException('FileVersion generating failed', $e);
-                }
-            } catch (\Exception $e) {
-                Log::error('FileVersion generating failed . ' . $e->getMessage());
-
-                return $this->responseException('Original File not readable', $e);
-            }
         }
-        if (!$imageWebP) {
-            return $this->response(['message' => 'FileVersion not found'], Response::HTTP_BAD_REQUEST);
+
+        $dataToUpdate = [
+            'name' => $imageName,
+            'extension' => $fileExt,
+            'path' => $path,
+            'full_path' => '/uploads' . $path,
+            'url' => Storage::disk('public')->url($path),
+            'folder_id' => $mediaFolder->id,
+            'parent_id' => $version == 'original' ? null : $request->input('parent_id'),
+            'alt_text' => $request->input('alt_text'),
+            'link' => $request->input('link'),
+            'size' => Storage::disk('public')->size($path),
+            'version' => $version,
+            'user_id' => Auth::user()->id,
+        ];
+
+        if ($imageWebP) {
+            $dataToUpdate += [
+                'height' => $imageWebP->height(),
+                'width' => $imageWebP->width(),
+                'crop_data' => $cropData,
+            ];
         }
+
 
         if (!$mediaFile) {
             $mediaFile = new MediaFile();
         }
-        $oldUrl = $mediaFile->url;
-        $mediaFile
-            ->fill([
-                'name' => $imageName,
-                'extension' => $fileExt,
-                'path' => $path,
-                'full_path' => '/uploads' . $path,
-                'url' => Storage::disk('public')->url($path),
-                'folder_id' => $mediaFolder->id,
-                'parent_id' => $version == 'original' ? null : $request->input('parent_id'),
-                'alt_text' => $request->input('alt_text'),
-                'link' => $request->input('link'),
-                'size' => Storage::disk('public')->size($path),
-                'height' => $imageWebP->height(),
-                'width' => $imageWebP->width(),
-                'crop_data' => $cropData ? json_encode($cropData) : null,
-                'version' => $version,
-                'user_id' => Auth::user()->id,
-            ])
-            ->save();
+
+        $mediaFile->fill($dataToUpdate)->save();
         $service->updateAltTextLink($mediaFile, $mediaFile->alt_text, $mediaFile->link);
 
-        if ($oldUrl) {
-            RenameFile::dispatch($oldUrl, $mediaFile->url, $mediaFile->alt_text, $mediaFile->link)
+        if ($mediaFile->url) {
+            RenameFile::dispatch($mediaFile->url, $mediaFile->url, $mediaFile->alt_text, $mediaFile->link)
                 ->delay(now()->addSeconds(5));
         }
